+++ src/main/java/com/dehold/contentmanager/validation/model/ValidationError.java
@@ -1,3 +1,3 @@
-package com.dehold.contentmanager.validation.result;
+package com.dehold.contentmanager.validation.model;
 
 public record ValidationError(String code, String message) {}
