+++ src/main/java/com/dehold/contentmanager/validation/model/ValidationResult.java
@@ -1,7 +1,6 @@
-package com.dehold.contentmanager.validation.result;
+package com.dehold.contentmanager.validation.model;
 
 import java.util.List;
-import java.util.Objects;
 
 public class ValidationResult {
     private final boolean isValid;
