+++ src/main/java/com/dehold/contentmanager/validation/pipeline/ValidationPipeline.java
@@ -1,9 +1,6 @@
 package com.dehold.contentmanager.validation.pipeline;
 
-import com.dehold.contentmanager.validation.result.ValidationResult;
-import com.dehold.contentmanager.validation.step.ValidationStep;
-
-import java.util.List;
+import com.dehold.contentmanager.validation.model.ValidationResult;
 
 public interface ValidationPipeline<T> {
     ValidationResult run(T content);
