+++ src/main/java/com/dehold/contentmanager/validation/step/LengthValidator.java
@@ -1,7 +1,7 @@
 package com.dehold.contentmanager.validation.step;
 
-import com.dehold.contentmanager.validation.result.ValidationError;
-import com.dehold.contentmanager.validation.result.ValidationResult;
+import com.dehold.contentmanager.validation.model.ValidationError;
+import com.dehold.contentmanager.validation.model.ValidationResult;
 
 import java.util.List;
 import java.util.function.Function;
