+++ src/main/java/com/dehold/contentmanager/validation/step/ValidationStep.java
@@ -1,6 +1,6 @@
 package com.dehold.contentmanager.validation.step;
 
-import com.dehold.contentmanager.validation.result.ValidationResult;
+import com.dehold.contentmanager.validation.model.ValidationResult;
 
 public interface ValidationStep<T> {
     ValidationResult validate(T content);
