+++ src/test/java/com/dehold/contentmanager/validation/pipeline/ValidationPipelineTest.java
@@ -1,7 +1,7 @@
 package com.dehold.contentmanager.validation.pipeline;
 
 import com.dehold.contentmanager.content.blogpost.model.BlogPost;
-import com.dehold.contentmanager.validation.result.ValidationResult;
+import com.dehold.contentmanager.validation.model.ValidationResult;
 import com.dehold.contentmanager.validation.step.LengthValidator;
 import com.dehold.contentmanager.validation.step.PhoneNumberForbiddenValidator;
 import com.dehold.contentmanager.validation.step.ValidationStep;
