+++ src/test/java/com/dehold/contentmanager/validation/step/PhoneNumberForbiddenValidatorTest.java
@@ -1,8 +1,8 @@
 package com.dehold.contentmanager.validation.step;
 
 import com.dehold.contentmanager.content.customersupport.model.SupportRequest;
-import com.dehold.contentmanager.validation.result.ValidationError;
-import com.dehold.contentmanager.validation.result.ValidationResult;
+import com.dehold.contentmanager.validation.model.ValidationError;
+import com.dehold.contentmanager.validation.model.ValidationResult;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
