+++ src/test/java/com/dehold/contentmanager/validation/web/ValidationControllerIntegrationTest.java
@@ -2,8 +2,8 @@ package com.dehold.contentmanager.validation.web;
 
 
 import com.dehold.contentmanager.content.blogpost.model.BlogPost;
-import com.dehold.contentmanager.validation.result.ValidationError;
-import com.dehold.contentmanager.validation.result.ValidationResult;
+import com.dehold.contentmanager.validation.model.ValidationError;
+import com.dehold.contentmanager.validation.model.ValidationResult;
 import com.dehold.contentmanager.validation.step.LengthValidator;
 import com.dehold.contentmanager.validation.web.dto.BlogPostValidationRequest;
 import com.dehold.contentmanager.validation.web.dto.ValidationResponse;
